requirements.txt
attrs==23.2.0
axial-positional-embedding==0.2.1
backcall==0.2.0
bert==2.2.0
boto3==1.26.79
botocore==1.29.79
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.28.3
cn-clip==1.5.1
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
docopt==0.6.2
einops==0.6.0
entrypoints==0.4
erlastic==2.0.0
fairscale==0.4.4
fastjsonschema==2.17.1
filelock==3.9.0
fonttools==4.38.0
fsspec==2024.2.0
ftfy==6.1.1
googletrans==3.0.0
graphviz==0.20.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.3
hyperframe==5.2.0
idna==2.10
imageio==2.31.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.6
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.0.0
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lit==17.0.6
lmdb==1.3.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
Pygments==2.17.2
pygtrans==1.5.3
pyparsing==3.1.1
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.9.0
pytorch-pretrained-bert==0.6.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
s3transfer==0.6.2
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
timm==0.9.16
tokenizers==0.15.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
widgetsnbextension==4.0.10
yacs==0.1.8
zipp==3.17.0
